Secure_amisp/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.30
mysql-connector-python
pydantic==2.7.1
python-dotenv==1.0.1
uvicorn==0.30.1
gunicorn==22.0.0